Updated task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int c;
-	long int n1, n2, fn;
+	int n1, n2, fn;
 
 	n1 = 1;
 	n2 = 2;
